Fix unwanted plugin trigger
If there was a `+` at the beginning of the line (ignoring whitespace prefix), this was enough for the plugin to trigger on several keybindings. This caused some unwanted operations in some cases. It is now mandatory to have a `---` or `===` right after the initial `+` to trigger the plugin. This complies with Emacs and grid style table syntaxes.